xmly_yhh/requirements.txt
django-redis==4.10.0
django-redis-cache==2.0.0
django-filter==2.1.0
django-cors-headers==3.0.2